% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_values.R
\name{assign_values}
\alias{assign_values}
\title{Assign environment variables to a list}
\usage{
assign_values(X, function_env, return_list)
}
\arguments{
\item{X}{\itemize{
\item the name of the variable in \code{function_env} to assign to the list slot named \code{X}
}}

\item{function_env}{\itemize{
\item the environment in which you want the function to search for the variable named \code{X}
}}

\item{return_list}{\itemize{
\item the list item you want to add the value of \code{X} to in the slot named \code{X}
}}
}
\description{
A function that may be called by \code{lapply()} to assign all of the variables in the given environment to slots within a list.
}
\seealso{
\code{\link[base:get]{base::get()}}
}
