% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasons.R
\name{create_seasons}
\alias{create_seasons}
\title{Create season attribute of the model}
\usage{
create_seasons(season_names, number_of_seasons, months_per_season)
}
\arguments{
\item{season_names}{\itemize{
\item a vector of season names
}}

\item{number_of_seasons}{\itemize{
\item the number of seasons
}}

\item{months_per_season}{\itemize{
\item a vector, where each object is the number of months in each season
}}
}
\value{
season_list a list object containing the seasons, where each season is a nested list with attributes id, name, and months for each season
}
\description{
A function returns a list with season attributes
}
