% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_assignment.R
\name{generic_assignment}
\alias{generic_assignment}
\title{Assign environment variables to a list}
\usage{
generic_assignment(list__, element_type, element_value)
}
\arguments{
\item{list__}{\itemize{
\item the list you want to affix objects to slots in
}}

\item{element_type}{\itemize{
\item a vector with the names of the slots you want the value to be assigned to in the list
}}

\item{element_value}{\itemize{
\item a vector with the values of the slots you are adding to the list
}}
}
\value{
the list with added slots
}
\description{
A function that creates a list with each element in the environment as a list slot
}
\seealso{
\code{\link[magrittr:inset2]{magrittr::inset2()}}
}
