% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_branch.R
\name{make_branch}
\alias{make_branch}
\title{Assign environment variables to a list}
\usage{
make_branch(
  analyst = "",
  study_name = "",
  movement_type = "",
  years = 1,
  seasons = 1,
  ages = 1,
  extended_plus_group = 1,
  first_year = 1,
  last_year = 1,
  season = 1,
  recruitment = "",
  growth = "",
  selectivity = "",
  movement = data.frame(1),
  natural_mortality = data.frame(1),
  fishing_mortality = data.frame(0.2),
  area = "",
  population = "",
  likelihood_component = "",
  fleet = "",
  survey = ""
)
}
\arguments{
\item{...}{each of the arguments needs to exist in your working directory. If they do, they will be assigned to a list that is then returned.}
}
\description{
A function that may be called by \code{lapply()} to assign all of the variables in the given environment to slots within a list. This
}
