% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{FUnction to make a data object from input list}
\usage{
make_data(config_object, data_object)
}
\arguments{
\item{config_object}{a list item generated from reading in a MAS JSON config}

\item{data_object}{a list item generated from reading in a MAS JSON data file}
}
\value{
a list with populated data objects
}
\description{
FUnction to make a data object from input list
}
