% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_obj.R
\name{make_data_obj}
\alias{make_data_obj}
\title{Function that creates a data object from a list}
\usage{
make_data_obj(
  data_list,
  name,
  units,
  data_object_type,
  sex,
  areas_of_operation,
  observation_error
)
}
\arguments{
\item{data_list}{the data values, a list which can contain multiple data entries}

\item{name}{a string denoting he name of the data source}

\item{units}{a string denoting the units of the data source}

\item{data_object_type}{the type of data (i.e. age composition, index)}

\item{sex}{the sex the data pertains to}

\item{areas_of_operation}{the areas this data source is collected in}

\item{observation_error}{the observation error associated with these data, a list which can contain multiple entries, each one is assumed to match an entry of data_list}
}
\value{
a list containing the data arguments
}
\description{
Function that creates a data object from a list
}
