% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_F.R
\name{make_F}
\alias{make_F}
\title{A function to make an F list}
\usage{
make_F(nyears, id = 1, estimated, delta_method, phase, values = NA)
}
\arguments{
\item{nyears}{The number of years in the model}

\item{id}{The ID of the time series, defaults to 1 assuming there is only one F time series.}

\item{estimated}{boolean, is F estimated or set?}

\item{delta_method}{boolean, is the delta method used to generate standard errors?}

\item{phase}{phase of F estimation}

\item{values}{a vector of length nyears that assigns the starting or fixed values for F}
}
\value{
the list of F properties and values
}
\description{
A function to make an F list
}
