% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_likelihood_component.R
\name{make_likelihood_component}
\alias{make_likelihood_component}
\title{Creates a likelihood component list for MAS}
\usage{
make_likelihood_component(ids, models, lambdas, parameters = NA, nyears)
}
\arguments{
\item{ids}{A vector of the IDs of each likelihood component in the model}

\item{models}{A vector with the model distributions for each likelihood type; allowable entries include "lognormal", "multinomial", and "dirichlet"}

\item{lambdas}{the weighting factors for each likelihood component, one for each year}

\item{parameters}{only needed for the dirichlet likelihoods, the vectors of observation error parameters}

\item{nyears}{number of years for each likelihood component}
}
\value{

}
\description{
Creates a likelihood component list for MAS
}
