% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list.R
\name{make_list}
\alias{make_list}
\title{Make a list of everything in function environment}
\usage{
make_list()
}
\value{
a list with the associated names and values of the environment objects
}
\description{
A function that returns a list representing the parent environment
}
