% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_SS_MAS.R
\name{map_SS_MAS}
\alias{map_SS_MAS}
\title{Function mapping SS values to MAS}
\usage{
map_SS_MAS(toMAS = TRUE, in_name)
}
\arguments{
\item{toMAS}{boolean; if this will be written to a MAS-style JSON input}

\item{in_name}{string; the string to map to a new name}
}
\value{
new_name - the mapped name
}
\description{
Function to map names between different stock assessment models
}
\author{
Christine Stawitz
}
