% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_matrix.R
\name{movement_matrix}
\alias{movement_matrix}
\title{Creates a movement matrix for MAS input}
\usage{
movement_matrix(recruits, males, females)
}
\arguments{
\item{recruits}{transition probability of recruits from one area to another}

\item{males}{transition probability of males from one area to another}

\item{females}{transition probability of females from one area to another}
}
\value{
the list with populated movement information
}
\description{
Creates a movement matrix for MAS input
}
