% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_lookup.R
\name{name_lookup}
\alias{name_lookup}
\title{Function to map stock assessment quantities between platforms}
\usage{
name_lookup(input_format, names, lookup)
}
\arguments{
\item{input_format}{string; the model the data or output is coming from}

\item{names}{string or vector of strings; names to map}

\item{lookup}{data frame that contains the lookup table to map in between different models}
}
\value{
new_name - the mapped name
}
\description{
Function to map names between different stock assessment models
}
\author{
Christine Stawitz
}
