% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plots.R
\name{output_plots}
\alias{output_plots}
\title{Function to read in a MAS output txt file (deprecated) and create plots.}
\usage{
output_plots(data.dir, ages, years, pop_name, rep_name, figs_dir)
}
\arguments{
\item{data.dir}{the directory where the output txt is stored}

\item{ages}{a vector of the ages in the model}

\item{years}{a vector of the years in the model}

\item{pop_name}{a string denoting the name of the population file}

\item{rep_name}{a string denoting the name of the report file}

\item{figs_dir}{the directory where output figures should be stored}
}
\value{
a list with two entries - parameter values and N at age
}
\description{
Function to read in a MAS output txt file (deprecated) and create plots.
}
