% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_N.R
\name{plot_NAtAge}
\alias{plot_NAtAge}
\title{A function to plot N at age}
\usage{
plot_NAtAge(SS, MAS, pdf_on = TRUE, byage = FALSE, years, ages)
}
\arguments{
\item{SS}{the matrix of numbers at age from Stock Synthesis, where rows are ages}

\item{MAS}{the matrix of numbers at age from MAS, where rows are ages}

\item{pdf_on}{boolean, should the PDF be created? Default is \code{TRUE}}

\item{byage}{boolean, should the numbers be plotted with each plot representing an age (default to FALSE) or a year?}

\item{years}{the vector of years in the model}

\item{ages}{the vector of ages in the model}
}
\description{
A function to plot N at age
}
