% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_config.R
\name{write_config}
\alias{write_config}
\title{This function writes a JSON configuration file for MAS.}
\usage{
write_config(config_list, outfile_path)
}
\arguments{
\item{config_list}{The list of objects, where each slot corresponds to a MAS config input}

\item{outfile_path}{The path to the config file you want to save. This should end in a .json file extension.}
}
\value{
a string comprising the JSON config file
}
\description{
This function writes a JSON configuration file for MAS.
}
