% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_test_data.R
\name{write_test_data}
\alias{write_test_data}
\title{This function writes dummy data that is used in the "Introduction" vignette}
\usage{
write_test_data(
  nyears,
  ewaa_vector = NULL,
  catch_index_v = NULL,
  survey_index_v = NULL,
  catch_comp_csv = NULL,
  survey_comp_csv = NULL
)
}
\arguments{
\item{nyears}{scalar, number of years in the model}

\item{ewaa_vector}{vector of empirical weight-at-age, length nyears * nages}

\item{catch_index_v}{list with two entries, one named "values" and one named "errors" corresponding to catch, each list entry has length nyears}

\item{survey_index_v}{list with two entries, one named "values" and one named "errors" corresponding to index, each list entry has length nyears}

\item{catch_comp_csv}{a string with the name of the .csv file containing the catch age composition data, if NULL (the default) dummy data are used}

\item{survey_comp_csv}{a string with the name of the .csv file containing the survey age composition data, if NULL (the default) dummy data are used}
}
\description{
This function writes dummy data that is used in the "Introduction" vignette
}
